feat: Exclude files, ANSI coloring

+++ macos/symlink-dotfiles.sh
@@ -2,7 +2,13 @@
 # Script to symlink dotfiles from a user-specified directory to $HOME
 # WARNING: This overwrites existing files, so make sure yours are backed up
 
-if [ $# -eq 0 ]; then
+local -a EXCLUDE_FILES=(
+  "README.md"
+  "secrets.sh.example"
+  ".DS_Store"
+)
+
+if [[ $# -eq 0 ]]; then
   echo "Usage: $0 <directory_name>"
   exit 1
 fi
@@ -11,15 +17,24 @@ DOTFILES_DIR="$PWD/$1"
 
 find "$DOTFILES_DIR" -type f | while read -r full_path; do
   file="${full_path#$DOTFILES_DIR/}"
+  filename=${file:t}
+
+  if [[ -n ${EXCLUDE_FILES[(r)$filename]} ]]; then
+    echo "\e[33mSkipping $file\e[0m"
+    continue
+  fi
 
   src_path="$DOTFILES_DIR/$file"
   dst_path="$HOME/$file"
 
+  dst_dir="$(dirname "$dst_path")"
+  mkdir -p "$dst_dir"
+
   rm "$dst_path" 2>/dev/null
 
   if ln -s "$src_path" "$dst_path"; then
-    echo "Symlinked $file"
+    echo "\e[32mSymlinked $file\e[0m"
   else
-    echo "Failed to symlink $file"
+    echo "\e[31mFailed to symlink $file\e[0m"
   fi
 done
